+++ projects/ng-rule-editor/src/lib/rule-editor.module.ts
@@ -1,21 +1,21 @@
 import { NgModule } from '@angular/core';
-import { RuleEditorComponent } from './rule-editor.component';
-
 import { FormsModule } from '@angular/forms';
+import { DragDropModule } from '@angular/cdk/drag-drop';
+import { ClipboardModule } from '@angular/cdk/clipboard';
+import { MatTooltipModule } from '@angular/material/tooltip';
+import { MatRadioModule } from '@angular/material/radio';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
 
+import { RuleEditorComponent } from './rule-editor.component';
 import { VariablesComponent } from './variables/variables.component';
 import { UneditableVariablesComponent } from './uneditable-variables/uneditable-variables.component';
 import { QuestionComponent } from './question/question.component';
-import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { DragDropModule } from '@angular/cdk/drag-drop';
 import { CalculateSumPromptComponent } from './calculate-sum-prompt/calculate-sum-prompt.component';
-import { MatRadioModule } from '@angular/material/radio';
 import { MathToFhirpathPipe } from './math-to-fhirpath.pipe';
 import { SyntaxConverterComponent } from './syntax-converter/syntax-converter.component';
 import { SyntaxPreviewComponent } from './syntax-preview/syntax-preview.component';
 import { QueryObservationComponent } from './query-observation/query-observation.component';
-
-
 
 @NgModule({
   declarations: [
@@ -33,10 +33,14 @@ import { QueryObservationComponent } from './query-observation/query-observation
     FormsModule,
     BrowserAnimationsModule,
     DragDropModule,
-    MatRadioModule
+    MatRadioModule,
+    ClipboardModule,
+    MatTooltipModule,
+    MatSnackBarModule
   ],
   exports: [
     RuleEditorComponent
   ]
 })
-export class RuleEditorModule { }
+export class RuleEditorModule {
+}