+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
+## [1.4.0] 2021-08-27
+### Added
+- Tooltip preview for automatically generated FHIRPath and x-fhir-query.
+- Copy button next to FHIRPath and x-fhir-query preview.
+
 ## [1.3.0] 2021-08-23
 ### Added
 - Add support for `x-fhir-query`.